+++ javaharness/java/arcs/android/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   package="arcs.android">
-
   <uses-sdk
     android:minSdkVersion="29"
     android:targetSdkVersion="29"/>
