+++ javaharness/java/arcs/android/demo/AutofillDemoActivity.java
@@ -12,7 +12,7 @@ import android.widget.TextView;
 
 import javax.inject.Inject;
 
-import arcs.android.ArcsAndroid;
+import arcs.android.ArcsAndroidClient;
 import arcs.api.PortableJsonParser;
 
 /** Autofill demo activity. Contains Autofill status info, and some example autofill fields. */
@@ -21,7 +21,7 @@ public class AutofillDemoActivity extends Activity {
   private static final int REQUEST_CODE_AUTOFILL_SET = 1;
 
   @Inject
-  ArcsAndroid arcs;
+  ArcsAndroidClient arcsAndroidClient;
 
   @Inject
   PortableJsonParser jsonParser;
@@ -44,12 +44,12 @@ public class AutofillDemoActivity extends Activity {
     Button capturePersonButton = findViewById(R.id.capture_person_button);
     capturePersonButton.setOnClickListener(v -> capturePerson());
 
-    arcs.connect(this);
+    arcsAndroidClient.connect(this);
   }
 
   @Override
   public void onDestroy() {
-    arcs.disconnect(this);
+    arcsAndroidClient.disconnect(this);
 
     super.onDestroy();
   }
@@ -86,7 +86,7 @@ public class AutofillDemoActivity extends Activity {
 
   private void openAutofillSettings() {
     Intent intent = new Intent(Settings.ACTION_REQUEST_SET_AUTOFILL_SERVICE);
-    intent.setData(Uri.parse("package:arcs.android.demo"));
+    intent.setData(Uri.parse("package:arcsAndroidClient.android.demo"));
     startActivityForResult(intent, REQUEST_CODE_AUTOFILL_SET);
   }
 
@@ -99,6 +99,6 @@ public class AutofillDemoActivity extends Activity {
     CapturePerson capturePerson = new CapturePerson();
     capturePerson.setId("capture-person-particle");
     capturePerson.setJsonParser(jsonParser);
-    arcs.runArc("IngestPeople", "capture-person-arc", "capture-person-pec", capturePerson);
+    arcsAndroidClient.runArc("IngestPeople", "capture-person-arc", "capture-person-pec", capturePerson);
   }
 }
