+++ javaharness/java/arcs/api/BUILD
@@ -14,7 +14,7 @@ android_library(
     ]),
     javacopts = ["-Xep:AndroidJdkLibsChecker:OFF"],
     deps = [
-        "//javaharness/java/arcs/crdt",
+        "//javaharness/java/arcs/crdt:crdt-android",
         "@com_google_dagger",
         "@javax_inject_source//jar",
         "@org_json//jar",
