+++ javaharness/java/arcs/api/UiRenderer.java
@@ -1,6 +1,5 @@
 package arcs.api;
 
 public interface UiRenderer {
-
   boolean render(PortableJson content);
 }
